Documentation (#5)
* updated README

* take out main version part before sending to function

* clean up

* updated

* moved to sbtsemverrelease package

* moving to sbtsemverrelease package

* updated imports

* updated test to demo customize auto bump

* simple refactoring

* split components based on numeric and non-numeric parts

* if "version.sbt" exists then no need to create temp file

* clean up

* setting snapshot flag

* updated README

+++ src/main/scala/sbtsemverrelease/SemanticVersioningReleasePlugin.scala
@@ -1,37 +1,30 @@
-package com.alphasystem.sbt.semver.release.plugin
+package sbtsemverrelease
 
 import com.alphasystem.sbt.semver.release._
+import com.alphasystem.sbt.semver.release.common.JGitAdapter
 import com.alphasystem.sbt.semver.release.internal._
 import sbt.Keys._
 import sbt._
 import sbtrelease.ReleasePlugin
-import sbtrelease.ReleasePlugin.autoImport.ReleaseTransformations._
 import sbtrelease.ReleasePlugin.autoImport._
+import sbtrelease.ReleaseStateTransformations._
 
 import java.io.File
-import scala.util.Properties
 import scala.util.matching.Regex
+import scala.util.{ Properties, Try }
 
 object SemanticVersioningReleasePlugin extends AutoPlugin {
 
   object autoImport {
-
-    type PreReleaseConfig =
-      com.alphasystem.sbt.semver.release.internal.PreReleaseConfig
 
     val determineVersion = taskKey[String](
       "A task to determine the next version to be without bumping actual version."
     )
 
     val startingVersion = settingKey[String](
-      """This option defines starting version of the build in case there is no tag available to determine next version. 
+      """This option defines the starting version of the build in case there is no tag available to determine next version. 
         |Default value is "0.1.0-SNAPSHOT".""".stripMargin.replaceNewLines
     )
-
-    val preReleaseStartingVersion =
-      settingKey[String](
-        "This option defines starting version of the pre-release. Default value is \"RC.1\"."
-      )
 
     val tagPrefix = settingKey[String](
       "This option defines prefix to use when tagging a release. Default value is \"v\"."
@@ -82,54 +75,21 @@ object SemanticVersioningReleasePlugin extends AutoPlugin {
         .replaceNewLines
     )
 
-    val preReleasePattern = settingKey[String](
-      """This option defines regular expression to filter tags based on pre-release pattern. Default value 
-        |is \".*+$\"""".stripMargin.replaceNewLines
-    )
-
     val preReleaseBump =
-      settingKey[(PreReleaseConfig, String) => String]("preReleaseBump")
-
-    val autoBumpMajorPattern = settingKey[Option[Regex]](
-      s"""This option defines the regular expression to bump "MAJOR" component. Default value is "\\[major]"."""
-    )
-
-    val autoBumpMinorPattern = settingKey[Option[Regex]](
-      s"""This option defines the regular expression to bump "MINOR" component. Default value is "\\[minor]"."""
-    )
-
-    val autoBumpPatchPattern = settingKey[Option[Regex]](
-      s"""This option defines the regular expression to bump "PATCH" component. Default value is "\\[patch]"."""
-    )
-
-    val autoBumpNewPreReleasePattern =
-      settingKey[Option[Regex]](
-        s"""This option defines the regular expression to bump "PRE_RELEASE" component. Default value is
-           | "\\[new-pre-release]".""".stripMargin.replaceNewLines
+      settingKey[(PreReleaseConfig, String) => String](
+        "This option defines the function that "
       )
 
-    val autoBumpPromoteToReleasePattern =
-      settingKey[Option[Regex]](
-        s"""This option defines the regular expression to bump a . Default value is "\\[promote]"."""
-      )
-
-    val majorVersionsMatching = settingKey[Int](
-      s"""This option defines filtering of a particular major version. Default value is "-1", which is 
-         |corresponds to no matching.""".stripMargin.replaceNewLines
+    val autoBump = settingKey[AutoBump](
+      "This option allows you to specify how the build version should be automatically bumped based on the contents of commit messages."
     )
 
-    val minorVersionsMatching = settingKey[Int](
-      s"""This option defines filtering of a particular minor version. Default value is "-1", which is 
-         |corresponds to no matching."""
-        .stripMargin
-        .replaceNewLines
+    val versionsMatching = settingKey[VersionsMatching](
+      "The option defines filtering of tags based on particular version."
     )
 
-    val patchVersionsMatching = settingKey[Int](
-      s"""This option defines filtering of a particular patch version. Default value is "-1", which is 
-         |corresponds to no matching."""
-        .stripMargin
-        .replaceNewLines
+    val preRelease = settingKey[PreReleaseConfig](
+      "This option defines configuration for pre-release"
     )
 
     object ReleaseKeys {
@@ -154,26 +114,13 @@ object SemanticVersioningReleasePlugin extends AutoPlugin {
         promoteToRelease = promoteToRelease.value,
         snapshot = snapshot.value,
         newPreRelease = newPreRelease.value,
-        autoBump = AutoBump(
-          autoBumpMajorPattern.value,
-          autoBumpMinorPattern.value,
-          autoBumpPatchPattern.value,
-          autoBumpNewPreReleasePattern.value,
-          autoBumpPromoteToReleasePattern.value
-        ),
-        versionsMatching = VersionsMatching(
-          majorVersionsMatching.value,
-          minorVersionsMatching.value,
-          patchVersionsMatching.value
-        ),
+        autoBump = autoBump.value,
+        versionsMatching = versionsMatching.value,
         componentToBump = componentToBump
           .value
           .map(VersionComponent.valueOf)
           .getOrElse(DefaultComponentToBump),
-        preReleaseConfig = PreReleaseConfig(
-          preReleaseStartingVersion.value,
-          preReleasePartPattern = preReleasePattern.value
-        ),
+        preReleaseConfig = preRelease.value,
         preReleaseBump = preReleaseBump.value
       )
     }
@@ -201,26 +148,44 @@ object SemanticVersioningReleasePlugin extends AutoPlugin {
       }
     }
 
+  private def initializeSnapshot = {
+    // JGitAdapter doesn't work well in tests where git is not initialized
+    Def.setting {
+      val computedValue =
+        Try(JGitAdapter(baseDirectory.value))
+          .toOption
+          .map(_.hasUncommittedChanges)
+          .getOrElse(DefaultSnapshot)
+      initializeSettingFromSystemProperty(
+        SnapshotSystemPropertyName,
+        computedValue
+      )
+    }
+  }
+
   private def initializeReleaseVersionFile =
     Def.setting {
-      val file = File.createTempFile("version_", ".sbt")
-      val ver =
-        if (releaseUseGlobalVersion.value)
-          (version in ThisBuild).value
-        else version.value
+      val defaultFile = new File(baseDirectory.value, "version.sbt")
+      if (defaultFile.exists()) defaultFile
+      else {
+        val file = File.createTempFile("version_", ".sbt")
+        val ver =
+          if (releaseUseGlobalVersion.value)
+            (version in ThisBuild).value
+          else version.value
 
-      file.deleteOnExit()
-      IO.write(
-        file,
-        s"""version in ThisBuild := "$ver""""
-      )
-      file
+        file.deleteOnExit()
+        IO.write(
+          file,
+          s"""version in ThisBuild := "$ver""""
+        )
+        file
+      }
     }
 
   override def projectSettings: Seq[Def.Setting[_]] = Seq[Setting[_]](
     releaseVersionFile := initializeReleaseVersionFile.value,
     startingVersion := getStartingVersion.value,
-    preReleaseStartingVersion := DefaultPreReleaseStartingVersion,
     tagPrefix := DefaultTagPrefix,
     tagPattern := DefaultTagPattern,
     snapshotSuffix := DefaultSnapshotSuffix,
@@ -232,10 +197,7 @@ object SemanticVersioningReleasePlugin extends AutoPlugin {
       PromoteToReleaseSystemPropertyName,
       DefaultPromoteToRelease
     ),
-    snapshot := initializeSettingFromSystemProperty(
-      SnapshotSystemPropertyName,
-      DefaultSnapshot
-    ),
+    snapshot := initializeSnapshot.value,
     newPreRelease := initializeSettingFromSystemProperty(
       NewPreReleaseSystemPropertyName,
       DefaultNewPreRelease
@@ -246,19 +208,11 @@ object SemanticVersioningReleasePlugin extends AutoPlugin {
         .getOrElse(DefaultComponentToBump.name())
     ),
     preReleaseBump := { (config: PreReleaseConfig, latestVersion: String) =>
-      DefaultPreReleaseBump(config, latestVersion)
+      defaultPreReleaseBump(config, latestVersion)
     },
-    preReleasePattern := ".*+$",
-    autoBumpMajorPattern := Some(AutoBump.DefaultMajorPattern),
-    autoBumpMinorPattern := Some(AutoBump.DefaultMinorPattern),
-    autoBumpPatchPattern := Some(AutoBump.DefaultPatchPattern),
-    autoBumpNewPreReleasePattern := Some(AutoBump.DefaultNewPreReleasePattern),
-    autoBumpPromoteToReleasePattern := Some(
-      AutoBump.DefaultPromoteToReleasePattern
-    ),
-    majorVersionsMatching := -1,
-    minorVersionsMatching := -1,
-    patchVersionsMatching := -1,
+    autoBump := AutoBump(),
+    versionsMatching := VersionsMatching(),
+    preRelease := PreReleaseConfig(),
     determineVersion :=
       SemanticBuildVersion(
         baseDirectory.value,
